CMS-1700: Organize functions
